`app/javascript/http_api/http.nonApiOnly(url)` is now a mirror of `app/javascript/http_api/api.apiOnly(url)`.
`app/javascript/components/search-bar/index.onSearch` called `http.post` with a dictionary `{url: action}` which caused an error.
Now it calls it with `action` as the url param (not inside a dictionary) and also with `skipJsonParsing: true` as an option,
so that `app/javascript/http_api/fetch.processResponse` doesn't try to Json-parse the html (which leads to another error).

`data` in `onSearch` is passed as `null` since on some views `fetch` can't process it, which leads to `400 bad response`.