Set ems storage "Region" label to provider_region
The 'description' instance method is not defined in the base
StorageManager class, or in
ManageIQ::Providers::IbmCloud::ObjectStorage::StorageManager.
As far as I can tell this is the only place it is used.

Setting the "Region" label directly to the 'provider_region' value
removes the dependency on defining a 'description' instance method (and
seems more relevant).