Convert assert_rbac from Rbac.search to Rbac.filtered
We are not using the attributes portion of Rbac.search()
So there is no need to calculate the record coutns

Also, passing in the ids will bring back all records and we will also
sort them to put them in the order of the ids passed in.

Converting to a sql count cuts the query down to only bringing back the ids
(if we fetch these records again later, then we may have incurred an extra query here)

Now, in the end, I would like to drop assert_rbac all together.
It is only used in a few places and they are relatively simple to remove,
but I didn't want to add the testing effort at this time.