Order Container#hourly_metrics for widgets
For metrics, when we have a where clause on timetamp we also order by timestamp.

Typically the records come back from the database in inserted order (read: ordered by timestamp),
so this tends to be in the right order.

A sporadic test found this issue:

spec/services/container_dashboard_service_spec.rb:460