Fix gtl so it displays icons for ext_management_system
COLUMN_WITH_IMAGE converts ext_management_system.name into an icon

Before
======

We currently pass the row record (vm or resource pool) to be decorated
ResourcePool is not totally based upon provider for the icons, so it sometimes breaks

After
=====

We pass in ext_management_system to be decorated for displaying ext_management_system records

Concerns
========

We converted the result rows into ruport objects. Then we lookup the result records again for other functionality.
With new lookup does not have all the includes and associations populated.
It is possibly that we do not have the ems populated in these new records. So this could introduce an n+1 looking up the ems.

Running it through a few screens did not show any increase in query count