Resolve ActionView::MissingTemplate errors by specifying template
Rails 7 now raises this error if you don't specify the template correctly.

Specifying templates with "." in the name was deprecated here:
https://www.github.com/rails/rails/pull/39164

The interface they want is without periods and for you to specify the formats
if it's non-standard.  Also, render with no :template, :partial, etc. is also
gone in rails 7.

Somewhat related issued with helpful information can be found here:
https://www.github.com/mileszs/wicked_pdf/issues/1005