Ownership class for controller parameter should be VmOrTemplate
It's was unclear at first why we don't hit this in rails 6.1
because on rails 7, we get VmOrTemplate as I would expect from this code:
https://github.com/ManageIQ/manageiq-ui-classic/blob/9761b105b3a0038f30c1b933c24bee99ea3daf0b/app/controllers/mixins/actions/vm_actions/ownership.rb#L71-L86

Keenan made the following findings:

This calls post 2 times: explorer and x_button.
In the explorer call, the db/model is ManageIQ::Providers::CloudManager::Vm
In the x_button the db/model is VmOrTemplate.

The explorer call sets the report_data_additional_options.model to the current model,
which is ManageIQ::Providers::CloudManager::Vm.

The x_button call will reuse the report_data_additional_options.model value if it is present.

In 6.1, the x_button has the instance variable set, so the model is ManageIQ::Providers::CloudManager::Vm
In 7.0, actionpack-7.0.8.4/lib/action_controller/metal/testing.rb has a method clear_instance_variables_between_requests.
So report_data_additional_options is not set and it sets the model to the current model, which is VmOrTemplate.