Use npmAuditExcludePackages instead of npmAuditIgnoreAdvisories
Ideally we could use npmAuditIgnoreAdvisories to be aware when new
advisories come in, but it turns out the IDs are not stable in that as
changes are made to a GHSA, it creates a new ID. This churn is difficult
to work with. Instead this commit switches to package based, which,
while not as granular, is more stable.