Fix issue when using kerberos auth with Rails 7
kerberos auth requires a wait_for_task under the covers. When we create
this task, we store a backup copy of the incoming request params within
the session object, so we can restore it later after the task we are
waiting for has completed, and that session object is `Marshal.dump`ed
into memcached. In Rails 7, Rails changed the ActionController::Parameters
object to include some more objects as context, such as the request, and
the request internally has an IO object that cannot be dumped. Since we
don't actually need the context and instead only need the parameters,
we can use `to_unsafe_h` instead of `deep_dup`, to get a copy of those
parameters in HashWithIndifferentAccess form. Later when wait_for_task
reconstructs the parameters, they will be placed back into the request
object, which will get them back into the ActionController::Parameters
format.