Add external auth spec
Follow up to #9295. If we revert that PR's changes, and run these tests
they fail when it tries to Marshal.dump the params internal IO object.

Note that the actual failure in test is slightly different from prod. In
test we get `TypeError: no _dump_data is defined for class StringIO`
because the TestRequest has a StringIO internally. In prod we get
`can't dump IO` because the real request has a real IO object
internally. Even so, this test still demonstrates the Marshal issue,
while also providing an end-to-end test for external auth login.