Drop j_str helper method
This method was added as a wrapper around the Rails `j` method to handle
non-string types. However, this feature was added in the Rails `j`
method as of Rails 6.0.0-rc1 in rails/rails@dd0cfb03, and so is
redundant.

This commit switches everything to the `j` method to be Rails idiomatic.