Disable miq_debug for cypress for now
This should allow us to remove a lot of the force: true we had to add
for many DOM elements near the top of the page as the notifications
could cover these elements and various cypress interactions with covered
elements would fail.

Note, we'll need to figure out how to test notifications differently.