Remove incorrect and unneeded assertions
We copy the dialog with the same description but different name.
We then select one of the two (copy or original) in the tree.
Either the copy or original is now displayed, the assertion on
the name could be wrong since we're asserting the name is the "copy"
name.

We're in the cleanup code.  We don't really care about these assertions.
We now select either of the two "Test Description" dialogs, remove the
first, verify the main div is correct, select the second and remove it.
There is no need to check the Name field or the CodeMirror-code to confirm
the removal was correct.