Support string interval coming from report preview
I hit this when running the UI tests without a simulate_queue_worker running.
I believe this is a bug that can happen where this code path provides an interval
of type string.

With no workers running, such as rails server with no simulate queue worker:
Go to: Overview->Reports->Reports accordion->Configuration->New Report

Give it a name, title, "base the report on" such as availability zone,
add a field such as Ems Ref to the selected fields.  Then, click "preview" tab.
Click "generate report preview" button... Before this change, it blows up with:

[TypeError] no implicit conversion of Integer into String

Note, this is a similar change as #9444