Cache login across specs
Seems to be cached for each spec in a file.  The next file will reset the local storage
and cookies so it appears to need to be done once per file.

Add validation to something that responds quickly..., /api

Previously, we'd land on /dashboard/show after login, now, visit
a faster page, /utilization after establishing or restoring session.