Extract the tenant id from the x_node
Using params for the action rbac_tenant_delete is correct for this action.

It should not be used for any role_allows? calls as they can be for doing
different actions/product features.  For example, after deleting a tenant,
the subsequent call to replace the right cell and update the screen after
the deletion should not refer to an already deleted tenant.

Fixes #9512