rexml 3.4.3 Document.new no longer accepts strings with no root element
Followup to https://github.com/ManageIQ/manageiq-gems-pending/pull/614

We generally create the xml document and add the root element and others afterwards.
This is ok, but we need to call REXML::Document.new(nil) instead.  In 3.4.3 and
future versions, it fails validation when the string provided to .new does not
contain a root element.

See the discussion in: https://www.github.com/ruby/rexml/pull/291