Rollback the tenant test changes due to timing issue failures
This commit reverts #9660, #9658 and part of #9633.

The database reset in the tenant spec is colliding with another puma thread
that is trying to do an API request at the same time.  We need to ensure the
reset is done in isolation to avoid this test failure that is happening more
than sporadically.